Refactored UIView tests to use Swift Testing
